feat: remove transitional apt package for snap (#6)
* chore: task naming standardized

* feat: transitional apt package for firefox snap removed

Removes the transitional package that is shipped with Ubuntu 24.04
that forces any installation of firefox via apt to install the package
via snap. Since the package has a specific version to it, this change
looks up the version name of the installed package and compares it to
the known value. If there is a match, it is purged.